float processing

+++ 2_aarch64_vs_armhf/inc/inc.h
@@ -4,9 +4,9 @@
 Process Y = M*X  + C
 */
 
-void lib_api_process_data_c(int* pX, int* pM, int* pC, int* pY, int len);
+void lib_api_process_data_c(float* pX, float* pM, float* pC, float* pY, int len);
 
 #if ARM_NEON == 1
-void lib_api_process_data_intrinsics(int* pX, int* pM, int* pC, int* pY, int len);
+void lib_api_process_data_intrinsics(float* pX, float* pM, float* pC, float* pY, int len);
 #endif
